requirements.txt
Django==2.0.9
django-cors-headers==2.4.0
djangorestframework==3.9.0
mysqlclient==1.3.14
names==0.3.0
pytz==2018.7
